refactor(magic): consolidate alt-guard helper, fix N+1 on offer listing
- Move _resolve_actor_sheet (and _resolve_endorser_sheet's logic) into
  services/auth.py so serializers no longer reach into views.py at
  runtime to break a circular import.
- _resolve_endorser_sheet stays as a one-line wrapper for callers
  using the old name; both helpers now share one query implementation.
- Cache the viewer's CharacterSheet list per-request in
  ThreadWeavingTeachingOfferSerializer to eliminate the N+1 that
  fired for every row in a paginated list response.
