+++ src/world/magic/views.py
@@ -19,7 +19,7 @@ from django.shortcuts import get_object_or_404
 from django.utils import timezone
 from django_filters.rest_framework import DjangoFilterBackend
 from evennia.accounts.models import AccountDB
-from rest_framework import mixins, serializers, status, viewsets
+from rest_framework import mixins, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.filters import OrderingFilter, SearchFilter
@@ -104,6 +104,7 @@ from world.magic.services import (
     preview_resonance_pull,
     resolve_pending_alteration,
 )
+from world.magic.services.auth import _resolve_endorser_sheet
 from world.magic.services.gain import account_for_sheet
 from world.roster.models import RosterEntry
 from world.stories.pagination import StandardResultsSetPagination
@@ -814,92 +815,10 @@ class ThreadWeavingTeachingOfferViewSet(viewsets.ReadOnlyModelViewSet):
 # =============================================================================
 
 # Error messages — module constants keep tests stable and satisfy STRING_LITERAL.
-_ERR_NO_ACTIVE_SHEET = "No active character sheet for this account."
-_ERR_ENDORSER_SHEET_REQUIRED = (
-    "endorser_sheet_id is required when account has multiple active tenures."
-)
-_ERR_ENDORSER_SHEET_INVALID = "Requested endorser_sheet_id is not among your active tenures."
 _ERR_ENDORSEMENT_SETTLED = "Endorsement already settled."
-_ERR_ACTOR_SHEET_REQUIRED = "{key} is required when account has multiple active tenures."
-_ERR_ACTOR_SHEET_INVALID = "Requested {key} is not among your active tenures."
 
-
-def _resolve_endorser_sheet(request: Request) -> CharacterSheet:
-    """Return the CharacterSheet to use as endorser for an incoming request.
-
-    Single active tenure → return that sheet.
-    Multiple active tenures → require explicit ``endorser_sheet_id`` in the POST
-    body (alt system guard — no implicit first-sheet selection per project
-    conventions).
-    No active tenures → raise PermissionDenied.
-
-    Shared by PoseEndorsementViewSet and SceneEntryEndorsementViewSet.
-    """
-    account = request.user
-    sheets = list(
-        CharacterSheet.objects.filter(
-            roster_entry__tenures__player_data__account=account,
-            roster_entry__tenures__end_date__isnull=True,
-        )
-    )
-    if not sheets:
-        raise PermissionDenied(_ERR_NO_ACTIVE_SHEET)
-    if len(sheets) == 1:
-        return sheets[0]
-    # Multiple active tenures — explicit sheet required.
-    requested_pk = request.data.get("endorser_sheet_id")  # noqa: STRING_LITERAL — HTTP request body key
-    if requested_pk is None:
-        raise serializers.ValidationError({"endorser_sheet_id": _ERR_ENDORSER_SHEET_REQUIRED})
-    try:
-        return next(s for s in sheets if s.pk == int(requested_pk))
-    except (StopIteration, ValueError) as exc:
-        raise PermissionDenied(_ERR_ENDORSER_SHEET_INVALID) from exc
-
-
-def _resolve_actor_sheet(
-    request: Request,
-    body_key: str,
-    *,
-    from_query: bool = False,
-) -> CharacterSheet:
-    """Return the CharacterSheet to use as the acting character for an incoming request.
-
-    Mirrors ``_resolve_endorser_sheet`` but accepts both POST body and GET query-param
-    sourcing via the ``from_query`` flag.  This is the shared alt-guard helper used by
-    any endpoint where the requesting account must pick a character to act as.
-
-    ``body_key`` names the request field that carries the explicit sheet PK:
-    - ``from_query=False`` (default) → reads from ``request.data`` (POST body)
-    - ``from_query=True`` → reads from ``request.query_params`` (GET endpoint)
-
-    Rules:
-    - No active tenures → raise ``PermissionDenied``.
-    - Single active tenure → return that sheet.
-    - Multiple tenures without explicit key → raise ``ValidationError``.
-    - Multiple tenures with valid key → return that sheet.
-    """
-    account = request.user
-    sheets = list(
-        CharacterSheet.objects.filter(
-            roster_entry__tenures__player_data__account=account,
-            roster_entry__tenures__end_date__isnull=True,
-        )
-    )
-    if not sheets:
-        raise PermissionDenied(_ERR_NO_ACTIVE_SHEET)
-    if len(sheets) == 1:
-        return sheets[0]
-    # Multiple active tenures — explicit sheet required.
-    source = request.query_params if from_query else request.data
-    requested_pk = source.get(body_key)  # noqa: STRING_LITERAL — parameterised body/query key
-    if requested_pk is None:
-        raise serializers.ValidationError(
-            {body_key: _ERR_ACTOR_SHEET_REQUIRED.format(key=body_key)}
-        )
-    try:
-        return next(s for s in sheets if s.pk == int(requested_pk))
-    except (StopIteration, ValueError) as exc:
-        raise PermissionDenied(_ERR_ACTOR_SHEET_INVALID.format(key=body_key)) from exc
+# _resolve_actor_sheet and _resolve_endorser_sheet are imported at the top of this module
+# from world.magic.services.auth — see import section above.
 
 
 class PoseEndorsementViewSet(
